+++ src/definitions.ts
@@ -80,9 +80,9 @@ export interface OAuth2AuthenticateBaseOptions {
      */
     resourceUrl?: string;
     /**
-     * PKCE is enabled by default when using @responseType 'code'. This options disables it if needed.
+     * Enable PKCE if you need it.
      */
-    pkceDisabled?: boolean;
+    pkceEnabled?: boolean;
     /**
      * A space-delimited list of permissions that identify the resources that your application could access on the user's behalf.
      * If you want to get a refresh token, you most likely will need the offline_access scope (only supported in Code Flow!)